delete on profile order

+++ pc-webshop/pages/api/deliveryinfo.js
@@ -25,8 +25,20 @@ export default async function handle(req, res) {
 
   if (method === "DELETE") {
     if (req.query?.id) {
+      const order = await Order.findOne({
+        _id: req.query.id,
+        userId: session.user.id,
+      });
+      if (!order) {
+        return res
+          .status(404)
+          .json({ error: "Order not found or unauthorized" });
+      }
+
       await Order.deleteOne({ _id: req.query.id });
-      res.json(true);
+      return res.json(true);
+    } else {
+      return res.status(400).json({ error: "Order ID not provided" });
     }
   }
 }
